Created new main class (Scrolls) to handle config reading, options etc. It will launch the ScrollsGenerator which handles the reporting.
