+++ build.gradle
@@ -7,7 +7,7 @@ version = '0.0.1-SNAPSHOT'
 
 description = """An Automated Delivery Report Engine"""
 
-mainClassName = 'se.diabol.scrolls.ScrollsGenerator'
+mainClassName = 'se.diabol.scrolls.Scrolls'
 
 sourceCompatibility = 1.8
 targetCompatibility = 1.8
