+++ src/test/groovy/se/diabol/scrolls/ScrollsGeneratorTest.groovy
@@ -1,31 +1,17 @@
 package se.diabol.scrolls
 
-import org.junit.runner.RunWith
-import org.spockframework.runtime.Sputnik
 import spock.lang.Specification
 
-import java.nio.charset.StandardCharsets
-
-@RunWith(Sputnik)
 class ScrollsGeneratorTest extends Specification {
 
-    private PrintStream outBefore
     def scrollsOutputFile = "build/Scrolls.html"
-
-    def setup() {
-        outBefore = System.out
-    }
-
-    def cleanup() {
-        System.setOut(outBefore)
-    }
 
     def "generate html report"() {
         setup: "Remove old release notes"
         "rm -f ${scrollsOutputFile}"
 
         when: "generating html report"
-        ScrollsGenerator generator = new ScrollsGenerator(config: []);
+        ScrollsGenerator generator = new ScrollsGenerator([], [], []);
         generator.generateHtmlReport(headerDataMock, reportsMock, scrollsOutputFile);
 
         then: "html report is generated without exceptions"
@@ -34,19 +20,6 @@ class ScrollsGeneratorTest extends Specification {
         assert htmlContent.contains("<td class=\"label\" class>New version:</td><td>V2-new</td>")
         assert htmlContent.contains("<p>Total 155 changes by 3 people, total 20 files</p>")
         assert htmlContent.contains("<p>2 Jira issues from 1 stories and 1 epics affected</p>")
-    }
-
-    def 'no scrolls generated when version options are missing'() {
-        final ByteArrayOutputStream sout = new ByteArrayOutputStream()
-        System.setOut(new PrintStream(sout))
-
-        //given: 'nothing'
-        when: 'no versions given'
-        ScrollsGenerator.invokeMethod('main', [].toArray())
-
-        then: "error is returned"
-        String out = sout.toString(StandardCharsets.UTF_8.name())
-        assert out.contains('error: Missing required options: old-version, new-version')
     }
 
     static def headerDataMock = [
